refactor: Update Account model to include email field
The `Account` model class in `me.project.todo.model` has been updated to include an `email` field. This field is used to store the email address associated with the account. The `Account` constructor has been modified to accept the `email` parameter, and getter and setter methods have been added for the `email` field. These changes allow for the storage and retrieval of email information for user accounts.